Discrete kernel function

Discrete kernel function

discrete_kernel(
  kernel = c("optimal", "triang", "epanech", "binomial"),
  x,
  z,
  h,
  k = NULL
)

Arguments

kernel

the type of kernel. Currently supported kernels are limited to: "optimal", "triang" and "epanech"

x

the target point at which the density is calculated

z

the vector of observations

h

the bandwidth (or smoothing parameter)

k

Optional: the integer (positive) parameter that defined the support of the kernel function (corresponds to parameter 'a' for triangular kernel). It is only used for optimal and triangular kernel

Value

Returns the value of the associated kernel function

See also

Examples

discrete_kernel(kernel = "optimal", x = 25, z = 1:50, h = 0.9, k = 20)
#>  [1] 0.00000000 0.00000000 0.00000000 0.00000000 0.01871809 0.01956892
#>  [7] 0.02037611 0.02113967 0.02185959 0.02253589 0.02316855 0.02375758
#> [13] 0.02430298 0.02480475 0.02526288 0.02567739 0.02604826 0.02637550
#> [19] 0.02665910 0.02689908 0.02709542 0.02724813 0.02735721 0.02742266
#> [25] 0.02744448 0.02742266 0.02735721 0.02724813 0.02709542 0.02689908
#> [31] 0.02665910 0.02637550 0.02604826 0.02567739 0.02526288 0.02480475
#> [37] 0.02430298 0.02375758 0.02316855 0.02253589 0.02185959 0.02113967
#> [43] 0.02037611 0.01956892 0.01871809 0.00000000 0.00000000 0.00000000
#> [49] 0.00000000 0.00000000
discrete_kernel(kernel = "triang", x = 25, z = 1:50, h = 10, k = 20)
#>  [1] 0.00000000 0.00000000 0.00000000 0.00000000 0.01013280 0.01659798
#>  [7] 0.02062695 0.02307282 0.02451486 0.02533755 0.02578975 0.02602797
#> [13] 0.02614746 0.02620407 0.02622915 0.02623939 0.02624319 0.02624444
#> [19] 0.02624479 0.02624487 0.02624488 0.02624488 0.02624488 0.02624488
#> [25] 0.02624488 0.02624488 0.02624488 0.02624488 0.02624488 0.02624487
#> [31] 0.02624479 0.02624444 0.02624319 0.02623939 0.02622915 0.02620407
#> [37] 0.02614746 0.02602797 0.02578975 0.02533755 0.02451486 0.02307282
#> [43] 0.02062695 0.01659798 0.01013280 0.00000000 0.00000000 0.00000000
#> [49] 0.00000000 0.00000000
discrete_kernel(kernel = "epanech", x = 25, z = 1:50, h = 20)
#>  [1] 0.000000000 0.000000000 0.000000000 0.000000000 0.000000000 0.003658537
#>  [7] 0.007129456 0.010412758 0.013508443 0.016416510 0.019136961 0.021669794
#> [13] 0.024015009 0.026172608 0.028142589 0.029924953 0.031519700 0.032926829
#> [19] 0.034146341 0.035178236 0.036022514 0.036679174 0.037148218 0.037429644
#> [25] 0.037523452 0.037429644 0.037148218 0.036679174 0.036022514 0.035178236
#> [31] 0.034146341 0.032926829 0.031519700 0.029924953 0.028142589 0.026172608
#> [37] 0.024015009 0.021669794 0.019136961 0.016416510 0.013508443 0.010412758
#> [43] 0.007129456 0.003658537 0.000000000 0.000000000 0.000000000 0.000000000
#> [49] 0.000000000 0.000000000
discrete_kernel(kernel = "binomial", x = 25, z = 1:50, h = 0.5)
#>  [1] 3.209642e-42 2.046147e-39 8.348279e-37 2.448133e-34 5.493610e-32
#>  [6] 9.806094e-30 1.428888e-27 1.730741e-25 1.765355e-23 1.530563e-21
#> [11] 1.135400e-19 7.238172e-18 3.975427e-16 1.882648e-14 7.681206e-13
#> [16] 2.693223e-11 8.079668e-10 2.060315e-08 4.424256e-07 7.897298e-06
#> [21] 1.150749e-04 1.333823e-03 1.183043e-02 7.541898e-02 3.077094e-01
#> [26] 6.035839e-01 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#> [31] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#> [36] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#> [41] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#> [46] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00