Fish dataset from the SIMTAP project

The SIMTAP project (Self-sufficient Integrated Multi-Trophic AquaPonic systems) aimed to develop sustainable aquaculture production system that, in particular, contributes to reduce fish feed inputs and resources consumption. We used data from an experiment in which gilthead seabream (sparus aurata) were stocked in 1.6 m3 tanks at a density of 1.5 kg·m-3 . Fish were reared for 46 days in a single recirculating aquaculture system composed of three rearing tanks. At the beginning of the experiment, a number n =200 of the fish were individually weighed (dg), and their length at the caudal fork (mm) was measured.

fish_data

Format

fish_data

A data frame with 200 rows and 2 columns:

length

Length in mm

weight

Weight in dg

...

Source

https://www.simtap.eu